+++ src/SIRTSolver.h
@@ -13,10 +13,10 @@ public:
 		const vector<double>& b = constraint->getRightSide();
 		const SparseMatrix& A = constraint->getMatrix();
 		double norm = 0.0;
-		for (int i = 0; i < A.size(); i++)
+		for (int i = 0; i < (int)A.size(); i++)
 			norm += constraint->getNorm(i);
 		//const SparseVector& columnCache = constraint->getColumnCache();
-		for (int i = 0; i < (int)b.size(); i++)
+		for (int i = 0; i < (int)(int)b.size(); i++)
 		{
 			double C = (b[i] - A[i].dot(x)) / norm;
 			for(auto it = A[i].cbegin(); it != A[i].cend(); it++)
@@ -28,10 +28,10 @@ public:
 		SparseVector db = A.transform(dx);
 		double lambda = 0.0;
 		double dbnorm = db.lengthSquared();
-		for (int i = 0; i < b.size(); i++)
+		for (int i = 0; i < (int)b.size(); i++)
 			lambda += db[i] * (b[i] - A[i].dot(x)) / dbnorm;
 			
-		for (int i = 0; i < (int)x.size(); i++)
+		for (int i = 0; i < (int)(int)x.size(); i++)
 		{
 			x[i] += lambda * weights[i] * dx[i];
 			dx[i] = 0.0;
@@ -40,7 +40,7 @@ public:
 
 	virtual void solve(PositiveConstraint* constraint)
 	{
-		for (int i = 0; i < x.size(); i++)
+		for (int i = 0; i < (int)x.size(); i++)
 			if (x[i] < 0.0) x[i] = 0.0;
 	}
 };
