+++ src/SparseMatrix.h
@@ -27,7 +27,7 @@ public:
 	SparseVector transform(const SparseVector& src) const
 	{
 		SparseVector result;
-		for(int i = 0; i < rows.size(); i++)
+		for (int i = 0; i < (int)rows.size(); i++)
 		{
 			result[i] = rows[i].dot(src);
 		}
@@ -37,7 +37,7 @@ public:
 	SparseVector transform(const vector<double>& src) const
 	{
 		SparseVector result;
-		for(int i = 0; i < rows.size(); i++)
+		for (int i = 0; i < (int)rows.size(); i++)
 		{
 			result[i] = rows[i].dot(src);
 		}
